refactor : 위치 변경, entity 구조 변경, 오타 수정 - error -> feign - user -> member - entity -> home/entity 로 entity 따로 빼기(erd 편하게 entity만 묶어서 관계를 편하게 보기 위해) - OneToMany로
