+++ backend/src/main/java/com/ssafy/home/domain/member/repository/LoginAttemptRepository.java
@@ -1,6 +1,6 @@
-package com.ssafy.home.domain.user.repository;
+package com.ssafy.home.domain.member.repository;
 
-import com.ssafy.home.domain.user.entity.LoginAttempt;
+import com.ssafy.home.entity.member.LoginAttempt;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface LoginAttemptRepository extends JpaRepository<LoginAttempt, Long> {
