+++ backend/src/main/java/com/ssafy/home/domain/member/repository/MemberRepository.java
@@ -1,6 +1,6 @@
-package com.ssafy.home.domain.user.repository;
+package com.ssafy.home.domain.member.repository;
 
-import com.ssafy.home.domain.user.entity.Member;
+import com.ssafy.home.entity.member.Member;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
