+++ backend/src/main/java/com/ssafy/home/domain/member/repository/MemberSecurityRepository.java
@@ -1,6 +1,6 @@
-package com.ssafy.home.domain.user.repository;
+package com.ssafy.home.domain.member.repository;
 
-import com.ssafy.home.domain.user.entity.MemberSecret;
+import com.ssafy.home.entity.member.MemberSecret;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface MemberSecurityRepository extends JpaRepository<MemberSecret, Long> {
