+++ backend/src/main/java/com/ssafy/home/domain/member/service/MemberService.java
@@ -1,16 +1,16 @@
-package com.ssafy.home.domain.user.service;
+package com.ssafy.home.domain.member.service;
 
-import com.ssafy.home.domain.user.dto.request.LoginRequest;
-import com.ssafy.home.domain.user.dto.request.RegisterRequest;
-import com.ssafy.home.domain.user.dto.response.TokenResponse;
-import com.ssafy.home.domain.user.entity.GeneralMember;
-import com.ssafy.home.domain.user.entity.LoginAttempt;
-import com.ssafy.home.domain.user.entity.Member;
-import com.ssafy.home.domain.user.entity.MemberSecret;
-import com.ssafy.home.domain.user.repository.GeneralMemberRepository;
-import com.ssafy.home.domain.user.repository.LoginAttemptRepository;
-import com.ssafy.home.domain.user.repository.MemberRepository;
-import com.ssafy.home.domain.user.repository.MemberSecurityRepository;
+import com.ssafy.home.domain.member.dto.request.LoginRequest;
+import com.ssafy.home.domain.member.dto.request.RegisterRequest;
+import com.ssafy.home.domain.member.dto.response.TokenResponse;
+import com.ssafy.home.domain.member.repository.GeneralMemberRepository;
+import com.ssafy.home.domain.member.repository.LoginAttemptRepository;
+import com.ssafy.home.domain.member.repository.MemberRepository;
+import com.ssafy.home.domain.member.repository.MemberSecurityRepository;
+import com.ssafy.home.entity.member.GeneralMember;
+import com.ssafy.home.entity.member.LoginAttempt;
+import com.ssafy.home.entity.member.Member;
+import com.ssafy.home.entity.member.MemberSecret;
 import com.ssafy.home.global.auth.Encryption;
 import com.ssafy.home.global.auth.jwt.JwtTokenProvider;
 import lombok.RequiredArgsConstructor;
