+++ backend/src/main/java/com/ssafy/home/domain/member/service/OauthLoginService.java
@@ -1,10 +1,12 @@
-package com.ssafy.home.domain.user.service;
+package com.ssafy.home.domain.member.service;
 
 
-import com.ssafy.home.domain.user.dto.OauthLoginDto;
-import com.ssafy.home.domain.user.entity.Member;
-import com.ssafy.home.domain.user.entity.MemberType;
-import com.ssafy.home.domain.user.repository.MemberRepository;
+import com.ssafy.home.domain.member.dto.OauthLoginDto;
+import com.ssafy.home.domain.member.repository.SocialMemberRepository;
+import com.ssafy.home.entity.member.Member;
+import com.ssafy.home.entity.member.MemberType;
+import com.ssafy.home.domain.member.repository.MemberRepository;
+import com.ssafy.home.entity.member.SocialMember;
 import com.ssafy.home.external.oauth.model.OAuthAttributes;
 import com.ssafy.home.external.oauth.service.SocialLoginApiService;
 import com.ssafy.home.external.oauth.service.SocialLoginApiServiceFactory;
@@ -25,6 +27,7 @@ public class OauthLoginService {
 
     private final JwtTokenProvider jwtTokenProvider;
     private final MemberRepository memberRepository;
+    private final SocialMemberRepository socialMemberRepository;
 
     public OauthLoginDto.Response oauthLogin(String accessToken, MemberType memberType) {
         SocialLoginApiService socialLoginApiService = SocialLoginApiServiceFactory.getSocialLoginApiService(memberType);
@@ -34,14 +37,27 @@ public class OauthLoginService {
         JwtTokenDto jwtTokenDto;
         Optional<Member> optionalMember = memberRepository.findByEmail(userInfo.getEmail());
         if(optionalMember.isEmpty()) { // 신규 회원 가입
-            Member oauthMember = userInfo.toMemberEntity(memberType);
+            Member oauthMember = userInfo.toMemberEntity();
             oauthMember = memberRepository.save(oauthMember);
+            socialMemberRepository.save(SocialMember.builder()
+                    .member(oauthMember)
+                    .memberType(memberType)
+                    .build());
 
             // 토큰 생성
             jwtTokenDto = jwtTokenProvider.createJwtTokenResponse(oauthMember);
             oauthMember.updateRefreshToken(jwtTokenDto.getRefreshToken());
         } else { // 기존 회원일 경우
             Member oauthMember = optionalMember.get();
+
+            // 해당 소셜 맴버 타입이 존재하지 않을 경우
+            Optional<SocialMember> optionalSocialMember = socialMemberRepository.findByMemberAndMemberType(oauthMember, memberType);
+            if(optionalSocialMember.isEmpty()){
+                socialMemberRepository.save(SocialMember.builder()
+                        .member(oauthMember)
+                        .memberType(memberType)
+                        .build());
+            }
 
             // 토큰 생성
             jwtTokenDto = jwtTokenProvider.createJwtTokenResponse(oauthMember);
