+++ backend/src/main/java/com/ssafy/home/global/auth/filter/JwtAuthenticationFilter.java
@@ -1,7 +1,7 @@
 package com.ssafy.home.global.auth.filter;
 
-import com.ssafy.home.domain.user.entity.Member;
-import com.ssafy.home.domain.user.repository.MemberRepository;
+import com.ssafy.home.entity.member.Member;
+import com.ssafy.home.domain.member.repository.MemberRepository;
 import com.ssafy.home.global.auth.jwt.JwtTokenProvider;
 import jakarta.servlet.FilterChain;
 import jakarta.servlet.ServletException;
