+++ backend/src/main/java/com/ssafy/home/global/auth/validator/OauthValidator.java
@@ -1,6 +1,6 @@
 package com.ssafy.home.global.auth.validator;
 
-import com.ssafy.home.domain.user.entity.MemberType;
+import com.ssafy.home.entity.member.MemberType;
 import com.ssafy.home.global.error.ErrorCode;
 import com.ssafy.home.global.error.exception.BusinessException;
 import org.springframework.stereotype.Service;
